requirements.txt
requests
beautifulsoup4
flask
python-dotenv
schedule
